+++ src/Neo/Persistence/ISnapshot.cs
@@ -9,7 +9,9 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
+using Neo.Json;
 using System;
+using System.Linq;
 
 namespace Neo.Persistence
 {
@@ -35,5 +37,40 @@ namespace Neo.Persistence
         /// <param name="key">The key of the entry.</param>
         /// <param name="value">The data of the entry.</param>
         void Put(byte[] key, byte[] value);
+
+        /// <summary>
+        /// Load data from json
+        ///
+        /// Expected data (in base64):
+        /// 
+        /// - "key":"value"
+        /// - "prefix": { "key":"value" }
+        /// </summary>
+        /// <param name="json">Json Object</param>
+        public void LoadFromJson(JObject json)
+        {
+            foreach (var entry in json.Properties)
+            {
+                if (entry.Value is JString str)
+                {
+                    // "key":"value" in base64
+
+                    Put(Convert.FromBase64String(entry.Key), Convert.FromBase64String(str.Value));
+                }
+                else if (entry.Value is JObject obj)
+                {
+                    // "prefix": { "key":"value" }  in base64
+
+                    foreach (var subEntry in obj.Properties)
+                    {
+                        if (subEntry.Value is JString subStr)
+                        {
+                            Put(Convert.FromBase64String(entry.Key).Concat(Convert.FromBase64String(subEntry.Key)).ToArray(),
+                                Convert.FromBase64String(subStr.Value));
+                        }
+                    }
+                }
+            }
+        }
     }
 }
